Add org link types for foreign language and quotations.
Language link type allows setting a language for export, and will display this
nicely in LaTeX output.

Quote link type will typeset the given content either inline or as a block
quote. Very useful for use in documents and formal writings, as it can
intelligently decide whether to insert as a block quote or inline.