Update org refile origin link to have file name as description
When refiling (or copying) an org element, instead of using the org heading name
as the description of the original location link (which will usually be the same
for the copied link), show the file name of the original location. The link
itself is unchanged, this is just the description, what is shown when the link
is folded.