Capture type buffer for getting quote text for org quote links.
Initial set up of a function to capture quote text, which can then be put in the
description of an org quote type link.

This uses a minor mode for setting up the buffer settings.

Some things to be added, like finding a way to quit the buffer rejecting the
edits (a la org-capture).

Also set up a method to edit a current link in this fashion. The function
defined here is generalised for that case, it can take an argument which is the
current text to be modified. But need to set up the system for getting that text
from the link.

Edited quote-link snippet for org mode to use this new set up.