Org-quote: only use buffer as helm-bibtex string if in notes directory
When selecting a source to use to insert a quotation as an org quote link, the
current buffer name was used as the default string for helm-bibtex completion.
This is very useful in notes buffers as it saves finding the current
book/article every time a quote is inserted. However, it is not useful when in
any other org buffer as it attempts to find a source whose bib key is the
current buffer name, which usually doesn't exist, and isn't the one that is
wanted anyway.

So check whether the current buffer is visiting a file in
`helm-bibtex-notes-path' directory, and if so, then use buffer name as default,
otherwise, do not have a default search.