Move enabling of `org-indent-mode` to main org config
Inheret auto-fill-mode and flyspell from text mode, so don't need to enable
those in org separately.
Also stop using lambda functions in hooks, as it is hard to remove them from
hooks, whereas if each feature for a hook is in an explicit, named function,
it's easy to enable or disable features of the hook.