Change default org-capture position for tasks
Instead of saving under the "Tasks" heading of todo.org, by default capture in
to "Inbox". This gives a more unified experience with capture from
elsewhere (e.g. shortcuts in iOS or using Drafts). It also means only one place
for captured content to be refiled from, and avoids having a nebulous heading
called Tasks under which are unorganised tasks.