Save external clipboard before killing
Anything saved into clipboard from an external application will  be added to the
kill ring before killing so that it is still accessible.

Very useful when copying something from an external app but realising I need to
do some clean up of the current location before pasting/yanking it in.