Org mode spelling set up
In org mode, there are easily identifiable parts of the buffer that I do not
want to be spell checked, because they contain non-English. This includes header
lines which define LaTeX export options, property drawers, link text and source
code blocks.

The one downside to this is that I have excluded all setup property lines from
spell check when using ispell/aspell. This includes the title line, which brings
the potential for spelling errors in the title of a document, which would make
me look very silly. However, this is mitigated by the fact that Flyspell is
still enabled, and so spelling errors are still underlined, just not brought to
my attention with ispell. This should be enough of a warning as I edit the title
to avoid any significant errors.

I could of course exclude the "#+TITLE:" line from the lines matched by
ispell-skip-region-alist, but because Emacs regexps don't support lookahead,
this would be quite complex, and doesn't seem worth the bother at this stage. If
it becomes a problem, I may revisit this later.