Create face for org-quote-links
Previously I had set org-quote links to have hard-coded face definition. The
problem with this is that when the colour theme of Emacs switched from light to
dark (as I like to have as day/night themes), the quote text became much less
readable in a dark theme if it were defined to suit a light theme. Now it has
its own face which inherits from org-date, which means that almost any colour
theme will set it to a sensible, readable, theme-fitting colour which stands out
about the right amount (by a very unscientific and subjective measure) from
surrounding text so as to be marked while also allowing continuous reading.