Disable quit on RETURN in multiple cursors.
I would rather use enter to do a new line than to exit all cursors. If I
accidentally press it, it's much easier to hit backspace than to re-select all
the cursor points again (especially if I have deleted whatever text was
initially marked, but not yet inserted the new copy).

Hitting enter happens not only when intending to enter a new line, but sometimes
in response to a question from or entry in the minibuffer -- hitting an extra
enter shouldn't cancel all multiple cursors. C-g is the natural way, to my mind,
of disabling multiple cursors, not ENTER.