Kill buffers to avoid user-interaction when publishing notes files
Because sitemap.org buffers were open, but being changed on disk during
org-publish routine, the user was required to confirm rereading from disk into
the buffer. This slows down publishing and makes Emacs hang until the user gives
the right choice. Removing this need allows publishing to be non-interactive,
which it should be.

This is achieved by killing any buffers visiting the sitemap.org files before
beginning the publishing regime.