Use autoload instead of require for themes.
Rather than requiring themes, use autoload for themes.

This improves startup time since I use `change-theme' to set a light and dark
theme depending on time of day. If both themes are required, then two themes
must be loaded before setting the theme. But if autoloaded, only the theme
appropriate to the current time of day is loaded, and the other theme is loaded
only when needed.