Auto-select helm-bibtex key when in a book note file.
Create function `adh-helm-bibtex` which calls `helm-bibtex`. If in a note in
book-notes directory, auto-select the resource which the current note refers to.
If in any other buffer, just use as normal (no pre-selection). If wishing to
select a different resource, can simply delete the default input from the
minibuffer prompt.