Defer loading of git-gutter+-mode until required.
Instead of loading and globally initialising git-gutter+-mode at init (which
takes significant time), defer loading until actually required.

"Actually required" means when visiting a file in a prog-mode derived mode or in
org-mode which is under version control. Any other file has no need for
git-gutter+-mode.

The enabling of the mode is handled by the relevant mode hooks. The loading is
handled by `autoload` and the function `enable-ggm-if-vc` is what is actually
run by the hooks in order to check if under version control.