Update .bib snippets to use 'date' rather than 'year'
Biblatex strongly recommends using the 'date' field instead of 'year'. Date is
general (can take a single year, as well as a range, a month, a specific date,
even a time, all generally handled in a single entry field). Year field
continues to be available, but really only for legacy use.

Along with some other small changes to the fields used for common entry types.