Save bookmarks when updated.
Emacs's default behaviour is that bookmarks are only saved when explicitly told
or when Emacs exits normally. If Emacs exists abnormally (is killed or crashes),
bookmark changes are lost.

A far more intuitive (to me) approach is to save bookmarks when updated. I
update bookmarks rarely, want changes to be permanent, and use registers for
temporary "bookmarks" which I don't want to keep around.